refactor: moved icons
