--- src/App.css
@@ -1,7 +1,0 @@
-.App {
-	text-align: center;
-}
-
-.App-header {
-	background-color: #222526;
-}
