--- src/App.tsx
@@ -1,10 +1,0 @@
-import './App.css';
-
-function App() {
-  return (
-    <div>
-    </div>
-  );
-}
-
-export default App;
