+++ src/axios/index.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
 
 export default axios.create({
-    baseURL: 'http://158.160.103.129:8080/v1/'
+    baseURL: 'http://localhost:8080/v1/'
 })
