+++ src/pages/AuthPage/AuthPage.tsx
@@ -8,7 +8,7 @@ import { fetchUser, login } from '../../store/actionCreators/actionCreatorsAuth'
 import { IUser } from '../../interfaces/user.interface';
 import styles from './AuthPage.module.scss';
 import TextField from '../../ui/TextField/TextField';
-import * as Icons from '../../icons/icons'
+import * as Icons from '../../shared/ui/icons'
 import Background from "../../ui/Background/Background";
 import {authSlice} from '../../store/slices/authSlice';
