+++ src/ui/QuestionHeader/QuestionHeader.tsx
@@ -1,6 +1,6 @@
 import {FC} from "react"
 import styles from "./QuestionHeader.module.scss"
-import * as Icons from "../../icons/icons"
+import * as Icons from "../../shared/ui/icons"
 
 const QuestionHeader: FC<{type: "question" | "task"}> = ({type}) =>
     <div className={styles["question-header"]}>
