+++ src/ui/QuestionSectionBottom/QuestionBottom.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react"
 import styles from "./QuestionBottom.module.scss"
 import Button from "../Button/Button"
-import * as Icons from "../../icons/icons"
+import * as Icons from "../../shared/ui/icons"
 import {getPluralAttempts, getPluralLeft, getPluralScore} from "../../utils/utils"
 
 interface QuestionBottomProps {
